app/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
PyMySQL==1.1.0
requests==2.32.4